+++ app/src/main/java/de/passwordvault/model/storage/Configuration.java
@@ -2,6 +2,9 @@ package de.passwordvault.model.storage;
 
 import android.content.Context;
 import android.content.SharedPreferences;
+
+import androidx.appcompat.app.AppCompatDelegate;
+
 import de.passwordvault.App;
 
 
@@ -10,7 +13,7 @@ import de.passwordvault.App;
  * to (unencrypted) shared preferences for configuration purposes.
  *
  * @author  Christian-2003
- * @version 3.5.0
+ * @version 3.5.1
  */
 public class Configuration {
 
@@ -30,9 +33,31 @@ public class Configuration {
     private static final String KEY_AUTOFILL_AUTHENTICATION = "autofill_authentication";
 
     /**
+     * Field stores the key with which to store whether to use dark- / light-mode.
+     */
+    private static final String KEY_APPEARANCE_DARKMODE = "darkmode";
+
+    /**
      * Field stores the shared preferences instance from which to retrieve data.
      */
     private static final SharedPreferences preferences = App.getContext().getSharedPreferences("settings", Context.MODE_PRIVATE);
+
+
+    /**
+     * Field stores a value indicating that the app uses light mode.
+     */
+    public static final int DARKMODE_LIGHT = 0;
+
+    /**
+     * Field stores a value indicating that the app uses dark mode.
+     */
+    public static final int DARKMODE_DARK = 1;
+
+    /**
+     * Field stores a value indicating that the app uses light- or dark mode depending on the system
+     * settings.
+     */
+    public static final int DARKMODE_SYSTEM = 2;
 
 
     /**
@@ -55,6 +80,7 @@ public class Configuration {
         return preferences.getBoolean(KEY_CONVERTED_STORAGE_1, false);
     }
 
+
     /**
      * Method changes whether the autofill service shall use caching.
      *
@@ -75,6 +101,7 @@ public class Configuration {
         return preferences.getBoolean(KEY_AUTOFILL_CACHING, true);
     }
 
+
     /**
      * Method changes whether the autofill service shall require authentication.
      *
@@ -93,6 +120,47 @@ public class Configuration {
      */
     public static boolean useAutofillAuthentication() {
         return preferences.getBoolean(KEY_AUTOFILL_AUTHENTICATION, true);
+    }
+
+
+    /**
+     * Method changes whether the app uses dark mode or light mode.
+     *
+     * @param setting   Whether the app uses dark or light mode.
+     */
+    public static void setDarkmode(int setting) {
+        SharedPreferences.Editor editor = preferences.edit();
+        editor.putInt(KEY_APPEARANCE_DARKMODE, setting);
+        editor.apply();
+    }
+
+    /**
+     * Method returns whether the app uses dark mode or light mode.
+     *
+     * @return  Whether the app uses dark or light mode.
+     */
+    public static int getDarkmode() {
+        return preferences.getInt(KEY_APPEARANCE_DARKMODE, DARKMODE_SYSTEM);
+    }
+
+
+    /**
+     * Method applies the darkmode setting to the app. This needs to be called whenever the darkmode
+     * changes and when the app starts.
+     */
+    public static void applyDarkmode() {
+        int darkmode = getDarkmode();
+        switch (darkmode) {
+            case DARKMODE_LIGHT:
+                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
+                break;
+            case DARKMODE_DARK:
+                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
+                break;
+            default:
+                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM);
+                break;
+        }
     }
 
 }